add correct axiosAddProduct
add correct axiosAddProduct
