+++ src/api/product.js
@@ -1,25 +1,12 @@
 import instance from './auth';
 
-export const axiosAddProduct = async userData => {
-  try {
-    const uploadPromise = instance.post('/product/add', userData.files, {
-      headers: {
-        'Content-Type': 'multipart/form-data',
-      },
-    });
-
-    const updatePromise = instance.post('/product/add', userData.data);
-
-    const [updateResponse] = await Promise.all([uploadPromise, updatePromise]);
-
-    console.log('Update Response:', updateResponse.data);
-    return {
-      updateResult: updateResponse.data,
-    };
-  } catch (error) {
-    console.error('Error:', error);
-    throw error;
-  }
+export const axiosAddProduct = async productData => {
+  const { data } = await instance.post('/product/add', productData, {
+    headers: {
+      'Content-Type': 'multipart/form-data',
+    },
+  });
+  return data;
 };
 
 // Delete product by Id
