+++ src/components/Shared/Button/Button.module.scss
@@ -2,8 +2,9 @@
   width: 196px;
   height: 50px;
   background: var(--background-btn-main-color);
-  box-shadow: 0px 3.43693px 3.43693px rgba(0, 0, 0, 0.25);
-  border-color: var(--background-btn-main-color);
+  // box-shadow: 0px 3.43693px 3.43693px rgba(0, 0, 0, 0.25);
+  // border-color: var(--background-btn-main-color);
+  outline: none;
   border-radius: 10.3108px;
   font-style: normal;
   font-weight: 500;
